pyproject.toml
[tool.poetry]
name = "smart-schedule-medical"
version = "0.1.0"
description = ""
authors = ["Christian Lopes <christian.lima@ntendencia.com.br>"]
readme = "README.md"
packages = [{include = "smart_schedule_medical"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
bcrypt = "^4.0.1"
asyncpg = "^0.28.0"
sqlalchemy2-stubs = "^0.0.2a35"
pydantic-sqlalchemy = "^0.0.9"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.9.1"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pytest = "^7.4.2"
pre-commit = "^3.4.0"
pytest-random-order = "^1.1.0"
pytest-cov = "^4.1.0"
aiosqlite = "^0.19.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
